+++ mesh_to_sdf/examples/demo.rs
@@ -12,7 +12,7 @@ fn main() {
     // easy_gltf use cgmath::Vector3 for vertices, which is compatible with mesh_to_sdf
     // you need to specify the cgmath feature in your Cargo.toml for mesh_to_sdf:
     // [dependencies]
-    // mesh_to_sdf = { version = "0.2.0", features = ["cgmath"] }
+    // mesh_to_sdf = { version = "0.2.1", features = ["cgmath"] }
     let vertices = vertices.iter().map(|v| v.position).collect_vec();
 
     // query points must be of the same type as vertices